requirements.txt
fastapi
uvicorn[standard]
numpy
weaviate-client
sentence-transformers
transformers
Pillow
tqdm
redis
faiss-gpu-cu12
faiss-cpu
python-dotenv
pydantic
jinja2
